+++ tests/testthat/helpers.R
@@ -1,9 +1,18 @@
+# a helper to sanitize the environment of quosures so that
+# expect_snapshot results are reproducible
+clear_quosure_environment <- function(x) {
+  if (rlang::is_quosure(x)) {
+    x <- rlang::quo_set_env(x, rlang::empty_env())
+  }
 
-# In some cases, the test value needs to be wrapped in an empty
-#  environment. If arguments are set in the model specification
-# (as opposed to being set by a `translate` function), they will
-# need this wrapper.
+  x
+}
 
-new_empty_quosure <- function(expr) {
-  new_quosure(expr, env = empty_env())
+# a helper to express the idiom of translating, subsetting out the
+# generated args, and snapshotting them
+translate_args <- function(x) {
+  x %>%
+    translate() %>%
+    purrr::pluck("method", "fit", "args") %>%
+    purrr::map(clear_quosure_environment)
 }